Performance optimizations. CardCollection now uses only 4 bytes of memory. Add test for card-related methods.

+++ src/Belot.Engine/Players/PlayCardAction.cs
@@ -16,6 +16,6 @@
 
         public PlayerPosition Player { get; internal set; }
 
-        public byte TrickNumber { get; internal set; }
+        public int TrickNumber { get; internal set; }
     }
 }
